fix(products): keep fractional 1C stock quantities exact
`stock[].quantity` and `stock[].reserve` were IntegerFields, so DRF cast
whatever 1C sent through int(). For anything sold by weight that silently
understated stock — 2.5 kg served as 2, and 0.5 kg served as 0, which
reads as out of stock and hides a sellable product outright.

Both are DecimalFields now, serialized as strings the way `price` in the
same nested serializer already is. That is a response shape change:
"3.000" where the API used to send 3. Every consumer of these two fields
already coerces with Number() — the balance row renderer, the add-to-cart
sheet, distributeStock, and the offline sync path — so the rendered
output is unchanged, and Number("3.000") is 3, not "3.000".

The order-item quantity on PurchaseOrderItem is a separate model field
and is untouched; only the live stock rows change.

distributeStock gains tests covering the string form. They are pinning an
existing contract rather than driving new code, so their teeth were
checked by mutation: a lexicographic sort in place of the numeric one
fails them, while dropping a Number() wrapper does not, because JS's `-`
coerces anyway.

+++ backend/core/serializers.py
@@ -529,8 +529,11 @@ class ProductSearchSerializer(serializers.Serializer):
     class StockSerializer(serializers.Serializer):
         warehouse = serializers.CharField(max_length=255)
         warehouse_name = serializers.CharField(max_length=255)
-        quantity = serializers.IntegerField()
-        reserve = serializers.IntegerField(read_only=True)
+        # 1C types these as Number and goods sold by weight come back
+        # fractional; an IntegerField floored 2.5 kg to 2 and 0.5 kg to 0.
+        # Decimal (serialized as a string, like `price` above) keeps them exact.
+        quantity = serializers.DecimalField(max_digits=15, decimal_places=3)
+        reserve = serializers.DecimalField(max_digits=15, decimal_places=3, read_only=True)
         price = serializers.DecimalField(max_digits=10, decimal_places=2, read_only=True)
 
     is_barcode = serializers.BooleanField(write_only=True)
